backend/requirements.txt

flask
mysql-connector-python
flask-cors
pillow
werkzeug